+++ src/test/java/com/xiaoleilu/hutool/demo/SystemUtilDemo.java
@@ -1,0 +1,9 @@
+package com.xiaoleilu.hutool.demo;
+
+import com.xiaoleilu.hutool.system.SystemUtil;
+
+public class SystemUtilDemo {
+	public static void main(String[] args) {
+		SystemUtil.dumpSystemInfo();
+	}
+}